# frontend/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Самый Крутой Бэк",
    "description": "Бэкэнд для управления договорами, счетами, платежами и шаблонами",
    "version": "1.0.0"
  },
  "paths": {
...
    "/templates/upload": {
      "post": {
        "tags": ["Templates"],
        "summary": "Upload Template",
        "description": "Загружает шаблон и сохраняет его в БД с динамическими полями.\n\n:param file: Загружаемый файл.\n:param template_type: Тип шаблона.\n:param display_name: Название шаблона.\n:param fields: Динамические поля.\n:param db: Сессия базы данных.\n:return: Подтверждение загрузки.",
        "operationId": "upload_template_templates_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_template_templates_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/templates/update": {
      "post": {
        "tags": ["Templates"],
        "summary": "Update Template",
        "description": "Обновляет динамические поля и display_name для шаблона.\n\n:param template_id: ID шаблона.\n:param display_name: Новое имя шаблона.\n:param fields: Новый список полей.\n:param db: Сессия базы данных.\n:return: Подтверждение обновления.",
        "operationId": "update_template_templates_update_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_template_templates_update_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_update_template_templates_update_post": {
        "properties": {
          "template_id": { "type": "integer", "title": "Template Id" },
          "display_name": { "type": "string", "title": "Display Name" },
          "fields": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": ["template_id", "display_name", "fields"],
        "title": "Body_update_template_templates_update_post"
      },
      "Body_upload_template_templates_upload_post": {
        "properties": {
          "file": { "type": "string", "format": "binary", "title": "File" },
          "template_type": { "type": "string", "title": "Template Type" },
          "display_name": { "type": "string", "title": "Display Name" },
          "fields": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": ["file", "template_type", "display_name", "fields"],
        "title": "Body_upload_template_templates_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}